WFD: W/A - Routing workaround for WFD.
Don't route to WFD device any sounds beside MUSIC stream
of MEDIA strategy.
WFD output is created once and couldn't be changed
depending on played media. So, if WFD output is
6-channels, any system sound couldn't be mixed as there
isn't any appropriate mixing implementation.
This patch avoids routing of system sounds to WFD for now.
It can be reverted as the appropriate usecases will be fully
implemented.

Change-Id: Ic1609b90709380ddfb8f5cf81266347ea511ddda
Signed-off-by: Sergii Iegorov <x0155539@ti.com>